+++ CHANGELOG.md
@@ -6,7 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.1.1] - 2018-11-24
+## [0.2.0] - 2018-11-25
+- [lukehsiao][lh]: Make dry-run the default and change --dry-run to --go.
+  Colorize output for clarity.
+
+## [0.1.1] - 2018-11-25
 - [lukehsiao][lh]: Improve docs.
 
 ## [0.1.0] - 2018-11-24